ui: rename port module as midi
